Current user added to comments

+++ app/controllers/comments_controller.rb
@@ -173,41 +173,49 @@ end
 # DELETE /comments/1
 def destroy
   @comment = Comment.find(params[:id])
-
-  if @comment.destroy
-    render json: { message: 'Comment deleted successfully' }, status: :ok
+  if @comment.user_id == @current_user.id
+    if @comment.destroy
+      render json: { message: 'Comment deleted successfully' }, status: :ok
+    else
+      render json: { errors: @comment.errors.full_messages }, status: :unprocessable_entity
+    end
   else
-    render json: { errors: @comment.errors.full_messages }, status: :unprocessable_entity
+    render json: { message: 'This user is not the creator' }, status: :unprocessable_entity
+
   end
+
 end
 
 
 # PUT /comments/1
   def update
     @comment = Comment.find(params[:id])
-
-    if @comment.update(comment_params)
-      render json: {
-        message: 'Comment updated successfully',
-        comment: {
-          id: @comment.id,
-          body: @comment.body,
-          post_id: @comment.post_id,
-          user_id: @comment.user_id,
-          created_at: @comment.created_at,
-          updated_at: @comment.updated_at,
-          num_replies: @comment.replies.count,
-          likes: {
-            positive: @comment.comment_likes.where(positive: true).count || 0,
-            negative: @comment.comment_likes.where(positive: false).count || 0
+    if @comment.user_id == @current_user.id
+      if @comment.update(comment_params)
+        render json: {
+          message: 'Comment updated successfully',
+          comment: {
+            id: @comment.id,
+            body: @comment.body,
+            post_id: @comment.post_id,
+            user_id: @comment.user_id,
+            created_at: @comment.created_at,
+            updated_at: @comment.updated_at,
+            num_replies: @comment.replies.count,
+            likes: {
+              positive: @comment.comment_likes.where(positive: true).count || 0,
+              negative: @comment.comment_likes.where(positive: false).count || 0
+            }
           }
-        }
-      }, status: :ok
-    else
-      render json: {
-        errors: @comment.errors.full_messages
-      }, status: :unprocessable_entity
-    end
+        }, status: :ok
+      else
+        render json: {
+          errors: @comment.errors.full_messages
+        }, status: :unprocessable_entity
+      end
+  else
+    render json: { message: 'This user is not the creator' }, status: :unprocessable_entity
+  end
   end
